CU-4hmj8e[closed]: optimize the function of declearing or revoking credential
